Logging

+++ modules/gui/include/gui/systems/parent_rectangle_changed_observer.h
@@ -14,24 +14,24 @@ namespace gui::systems {
     inline void on_parent_rectangle_changed_observer(flecs::entity e, const Rectangle &parent, Rectangle &rec, const Anchor& anchor) {
         switch (anchor.horizontal_anchor) {
             case CENTER:
-                rec.x = anchor.position.x + parent.x + parent.width / 2;
+                rec.x = (int)(anchor.position.x + parent.x + parent.width / 2);
                 break;
             case RIGHT:
-                rec.x = anchor.position.x + parent.x + parent.width;
+                rec.x = (int)(anchor.position.x + parent.x + parent.width);
                 break;
             default:
-                rec.x = anchor.position.x + parent.x;
+                rec.x = (int)(anchor.position.x + parent.x);
                 break;
         }
         switch (anchor.vertical_anchor) {
             case MIDDLE:
-                rec.y = anchor.position.y + parent.y + parent.height / 2;
+                rec.y = (int)(anchor.position.y + parent.y + parent.height / 2);
                 break;
             case BOTTOM:
-                rec.y = anchor.position.y + parent.y + parent.height;
+                rec.y = (int)(anchor.position.y + parent.y + parent.height);
                 break;
             default:
-                rec.y = anchor.position.y + parent.y;
+                rec.y = (int)(anchor.position.y + parent.y);
                 break;
         }
